Fix Result Parsing
This fixes a couple of probems with the results parsing. When the summit height was reported as ‘-’ the regex match would fail and therefore the summit height was shown as null. Some resorts lifts are reported as 0 when all lifts are closed and some just lack the data. Inserting a 0 in case of missing data fixes this. Also all heights are now always displayed as a number instead of - when there is no snow.